requirements.txt
# Core Data Science & ML
pandas
numpy
torch
faiss-cpu

# Hugging Face & RAG Libraries
transformers
sentence-transformers
langchain

# Evaluation
ragas
datasets

# Application & UI
streamlit
python-docx
fpdf2

# Utilities
PyYAML
psutil
tqdm
ipykernel